better parsing of value and parameter
The parameter is now considered implicit rather than explicit and the
next rule (rules[i+1], if such index exists) is always sent to the
cleanVal() method as the parameter for the current rule. The method
decides for itself wether to use the parameter or not. This way it's
also possible to implement validation rules that use optional
parameters (to be explored in future releases).